Compiles again

+++ src/ModuleRam.cpp
@@ -17,7 +17,7 @@ void ModuleRam::draw() {
 			for(; i != snapshots->end();) {
 				if(*i && *i != nullptr) {
 					if(last == nullptr || (last->getAge() - (*i)->getAge()).asSeconds() < 3) {
-						points.push_back(new sf::Vector2f((1 - (*i)->getAge().asSeconds() / 180F) * getDisplayWidth(),
+						points.push_back(new sf::Vector2f((1 - (*i)->getAge().asSeconds() / (float) 180) * getDisplayWidth(),
 														  (1 - (float) (*i)->getMem() / (*i)->getTotalmem()) *
 														  getDisplayHeight()));
 						last = *i;
